Initial setup
This is a reproduction of the app used for the test, with the serial numbers filled out, and updated to Rails 7.1 instead of 6.0. But the basics are the same: a new Rails API-only application (see https://guides.rubyonrails.org/api_app.html for more details). The only thing missing from the original tests are the RSpec files.